Sort some data on-the-fly instead of pre-sorting.
This is a little slower, but it saves us some memory. Most importantly,
it paves the way to when we can get rid of the pre-sorted `Bok` matrix
which is a pretty large memory hog. A first attempt to do this did not
work too well, so we will try again later.